third_party/blink/web_tests/platform/generic/fast/events/popup-blocking-timers4-expected.txt

Click Here
PASS successfullyParsed is true

TEST COMPLETE

Test calling window.open() in a nested call to setTimeout(). A popup should be allowed.
PASS newWindow is non-null.
